MemeComponent isolation
