+++ package.json
@@ -27,6 +27,7 @@
     "file-loader": "^4.3.0",
     "html-webpack-plugin": "^3.2.0",
     "jest": "^24.9.0",
+    "react-image-resizer": "^1.3.0",
     "style-loader": "^1.0.2",
     "webpack": "^4.41.4",
     "webpack-cli": "^3.3.10",
@@ -40,6 +41,7 @@
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-image-autofit-frame": "^0.1.0",
+    "react-resize-image": "^1.0.4",
     "react-router-dom": "^5.1.2",
     "react-vk": "^5.0.2",
     "recharts": "^1.8.5"
